build/init.common: prevent arp responses on the wrong nic.
Having multiple NICs in the same segment that are going to be bonded
is a common scenario in modern datacenters. The kernel should not
respond to arp requests for ip addresses that are not for the
specific NIC it arrives on.

This patch prevents dhclient from dropping its ip address while
renewing its dhcp lease. Closes #237